Api definitions for Joson and Bson in Java
