+++ src/main/java/io/boson/bson/BsonObject.java
@@ -43,6 +43,7 @@ import static java.time.format.DateTimeFormatter.ISO_INSTANT;
  *
  * @author <a href="http://tfox.org">Tim Fox</a>
  */
+@Deprecated
 public class BsonObject implements Iterable<Map.Entry<String, Object>> {
 
     private Map<String, Object> map;
