--- src/main/java/io/boson/json/JsonExtractorBuilder.java
@@ -1,15 +1,0 @@
-package io.boson.json;
-
-
-import io.boson.ExtractorBuilder;
-
-import javax.json.Json;
-import java.util.ArrayList;
-import java.util.List;
-
-
-public class JsonExtractorBuilder<I>  {
-
-   // public <O> JsonExtractor<O>
-
- }
