+++ src/main/java/io/boson/json/extractors/ExtractorBuilder.java
@@ -1,4 +1,4 @@
-package io.boson;
+package io.boson.json.extractors;
 
 public interface ExtractorBuilder<T> {
