+++ src/main/java/io/boson/json/extractors/JsonExtractorBuilder.java
@@ -1,0 +1,8 @@
+package io.boson.json.extractors;
+
+
+public class JsonExtractorBuilder<I>  {
+
+   // public <O> JsonExtractor<O>
+
+ }
