+++ src/test/java/io/boson/BuilderTest.java
@@ -1,10 +1,7 @@
 package io.boson;
 
 
-import io.boson.json.JsonExtractor;
 
-import io.boson.json.ObjectExtractor;
-import io.boson.json.StringExtractor;
 import org.junit.Test;
 
 import javax.json.Json;
@@ -12,7 +9,6 @@ import javax.json.stream.JsonParser;
 import java.io.StringReader;
 
 
-import static org.junit.Assert.assertEquals;
 
 
 public class BuilderTest {
@@ -37,11 +33,11 @@ public class BuilderTest {
     public void buildExtractors() {
         JsonParser parser = Json.createParser(new StringReader(json));
 
-        JsonExtractor<String> ext  = new ObjectExtractor<>(new StringExtractor("onclick") );
-
-        String result = ext.apply(parser).getResult().toString();
-
-        assertEquals(result, "CreateNewDoc()" );
+//        JsonExtractor<String> ext  = new ObjectExtractor<>(new StringExtractor("onclick") );
+//
+//        String result = ext.apply(parser).getResult().toString();
+//
+//        assertEquals(result, "CreateNewDoc()" );
 
         // to get the OpenDoc() method name
 //        Extractor ext = BuildExtractor
